+++ backend/Onied/Storage/Storage/Abstractions/IStorageService.cs
@@ -1,0 +1,7 @@
+namespace Storage.Abstractions;
+
+public interface IStorageService
+{
+    Task<IResult> Upload(IFormFileCollection files);
+    Task<IResult> Get(string objectName);
+}
