+++ backend/Onied/Storage/Storage/Commands/Upload.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace Storage.Commands;
+
+public record Upload(IFormFileCollection Files) : IRequest<IResult>;
