+++ backend/Onied/Storage/Storage/Handlers/GetHandler.cs
@@ -1,0 +1,13 @@
+using MediatR;
+using Storage.Abstractions;
+using Storage.Queries;
+
+namespace Storage.Handlers;
+
+public class GetHandler(IStorageService storageService) : IRequestHandler<Get, IResult>
+{
+    public async Task<IResult> Handle(Get request, CancellationToken cancellationToken)
+    {
+        return await storageService.Get(request.ObjectName);
+    }
+}
