+++ backend/Onied/Storage/Storage/Program.cs
@@ -1,8 +1,11 @@
 using Minio;
+using Storage.Abstractions;
 using Storage.Middlewares;
+using Storage.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
+builder.Services.AddScoped<IStorageService, StorageService>();
 
 builder.Services.AddSingleton<ExceptionMiddleware>();
 
@@ -12,6 +15,8 @@ builder.Services.AddMinio(configureClient => configureClient
     .WithCredentials(minioConfiguration["AccessKey"], minioConfiguration["SecretKey"])
     .WithSSL(false)
     .Build());
+
+builder.Services.AddMediatR(configuration => configuration.RegisterServicesFromAssemblyContaining<Program>());
 
 // Add services to the container.
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
