+++ backend/Onied/Storage/Storage/Queries/Get.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace Storage.Queries;
+
+public record Get(string ObjectName) : IRequest<IResult>;
